# proto/gen/statsparrot/runtime/v1/runtime.swagger.yaml
swagger: "2.0"
info:
  title: statsparrot/runtime/v1/schema.proto
  version: version not set
...
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /v1/instances:
    get:
      summary: ListInstances lists all the instances currently managed by the runtime
      operationId: RuntimeService_ListInstances
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1ListInstancesResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: pageSize
          in: query
          required: false
          type: integer
          format: int64
        - name: pageToken
          in: query
          required: false
          type: string
      tags:
        - RuntimeService
    post:
      summary: CreateInstance creates a new instance
      operationId: RuntimeService_CreateInstance
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1CreateInstanceResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          description: |-
            Request message for RuntimeService.CreateInstance.
            See message Instance for field descriptions.
          in: body
          required: true
          schema:
            $ref: '#/definitions/v1CreateInstanceRequest'
      tags:
        - RuntimeService
...
definitions:
...
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
...
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
...
  v1Connector:
    type: object
    properties:
      type:
        type: string
        description: Type of the connector. One of the infra driver supported.
      name:
        type: string
        title: Name of the connector
      config:
        type: object
        title: Config for the connector
      templatedProperties:
        type: array
        items:
          type: string
        title: Properties in config that use templating
      provision:
        type: boolean
        title: Provision settings if its a managed connector
      provisionArgs:
        type: object
...
  v1CreateInstanceRequest:
    type: object
    properties:
      instanceId:
        type: string
      environment:
        type: string
      olapConnector:
        type: string
      repoConnector:
        type: string
      adminConnector:
        type: string
      aiConnector:
        type: string
      connectors:
        type: array
        items:
          type: object
          $ref: '#/definitions/v1Connector'
      variables:
        type: object
        additionalProperties:
          type: string
        description: 'Deprecated: Continue populating this for some time for backward compatibility with older runtimes, remove after a few releases.'
      systemVariables:
        type: object
        additionalProperties:
          type: string
        title: system_variables are variables that are set by admin service and should not be overridden by user input
      annotations:
        type: object
        additionalProperties:
          type: string
      frontendUrl:
        type: string
    description: |-
      Request message for RuntimeService.CreateInstance.
      See message Instance for field descriptions.
  v1CreateInstanceResponse:
    type: object
    properties:
      instance:
        $ref: '#/definitions/v1Instance'
    title: Response message for RuntimeService.CreateInstance
...
  v1Instance:
    type: object
    properties:
      instanceId:
        type: string
      environment:
        type: string
      projectDisplayName:
        type: string
      olapConnector:
        type: string
      repoConnector:
        type: string
      adminConnector:
        type: string
      aiConnector:
        type: string
      createdOn:
        type: string
        format: date-time
      updatedOn:
        type: string
        format: date-time
      connectors:
        type: array
        items:
          type: object
          $ref: '#/definitions/v1Connector'
      projectConnectors:
        type: array
        items:
          type: object
          $ref: '#/definitions/v1Connector'
      variables:
        type: object
        additionalProperties:
          type: string
      projectVariables:
        type: object
        additionalProperties:
          type: string
      featureFlags:
        type: object
        additionalProperties:
          type: boolean
      annotations:
        type: object
        additionalProperties:
          type: string
      aiInstructions:
        type: string
      frontendUrl:
        type: string
      theme:
        type: string
    description: |-
      Instance represents a single data project, meaning one set of code artifacts,
      one connection to an OLAP datastore (DuckDB, Druid), and one catalog of related
      metadata (such as reconciliation state). Instances are the unit of isolation within
      the runtime. They enable one runtime deployment to serve not only multiple data
      projects, but also multiple tenants. On local, the runtime will usually have
      just a single instance.
...
  v1ListInstancesResponse:
    type: object
    properties:
      instances:
        type: array
        items:
          type: object
          $ref: '#/definitions/v1Instance'
      nextPageToken:
        type: string
    title: Response message for RuntimeService.ListInstances
...